feat(core): nx list enhancements and cleanup
