+++ packages/workspace/src/utils/plugins/index.ts
@@ -1,0 +1,13 @@
+export {
+  fetchCommunityPlugins,
+  listCommunityPlugins
+} from './community-plugins';
+export { fetchCorePlugins, listCorePlugins } from './core-plugins';
+export {
+  getInstalledPluginsFromNodeModules,
+  listInstalledPlugins
+} from './installed-plugins';
+export {
+  getPluginCapabilities,
+  listPluginCapabilities
+} from './plugin-capabilities';
