+++ packages/workspace/src/utils/plugins/shared.ts
@@ -1,0 +1,5 @@
+// Lifted in part from https://github.com/nrwl/angular-console
+
+export function hasElements(obj: any): boolean {
+  return obj && Object.values(obj).length > 0;
+}
